Only install python2 if not already there

+++ tools/build/node-debian.sh
@@ -4,8 +4,10 @@ set -e
 source $NVM_DIR/nvm.sh
 
 # ensure python2 is available by name for gyp
-apt-get update && apt-get install -y python2.7 --no-install-recommends
-ln -s /usr/bin/python2.7 /usr/bin/python2
+if [ ! -f /usr/bin/python2 ]; then
+  apt-get update && apt-get install -y python2.7 --no-install-recommends
+  ln -sn /usr/bin/python2.7 /usr/bin/python2
+fi
 
 nvm install $NODE_VERSION
 nvm alias default $NODE_VERSION
